Initial Commit
This example plug-in solution will prevent emails from being automatically promoted to Dynamics in the event that an email is attempted to be created due to in-reply-to correlation, but the recipients of the item do not match the original item being replied to. This behavior is most commonly seen when an email is forwarded to a different system user, and is unintentionally promoted to Dynamics.